% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcon22.R
\name{a_count_occurrences_trtem_ae}
\alias{a_count_occurrences_trtem_ae}
\title{Count occurrences of a variable in a dataframe}
\usage{
a_count_occurrences_trtem_ae(
  df,
  .var,
  .N_col,
  df_denom = NULL,
  denom = c("N_s", "N_col", "n"),
  id_var = "USUBJID",
  arm_var = "ARM"
)
}
\arguments{
\item{df}{The input dataframe.}

\item{.var}{The variable to count occurrences.}

\item{.N_col}{The column to count. This parameter is ignored by lintr.}

\item{df_denom}{An optional dataframe for denominator. Default is NULL.}

\item{denom}{A character vector for denominator. Default is c("N_s", "N_col", "n").}

\item{id_var}{The variable for id. Default is "USUBJID".}

\item{arm_var}{The variable for arm. Default is "ARM".}
}
\value{
A dataframe with the count and proportion of each occurrence of the variable.
}
\description{
This function counts the occurrences of a variable in a dataframe, and calculates the proportion of each occurrence.
}
\examples{
\dontrun{
a_count_occurrences_trtem_ae(df, .var = "AGEGR1")
}
}
