% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcon03.R
\name{create_sassy_report_03}
\alias{create_sassy_report_03}
\title{Create Sassy Report}
\usage{
create_sassy_report_03(
  df_tbl,
  titles,
  footnotes,
  page_header,
  page_footer,
  headers,
  column_labels,
  file = "fda-table-03.rtf"
)
}
\arguments{
\item{df_tbl}{The input dataframe.}

\item{titles}{The titles for the report.}

\item{footnotes}{The footnotes for the report.}

\item{page_header}{The header for each page.}

\item{page_footer}{The footer for each page.}

\item{headers}{The headers for the spanning columns.}

\item{column_labels}{The labels for the columns.}

\item{file}{The file name for the report. Default is "fda-table-22.rtf".}
}
\value{
The path of the created report.
}
\description{
This function creates a sassy report based on the input dataframe.
}
\examples{
\dontrun{
report_path <- create_sassy_report(
  df_tbl,
  headers,
  titles,
  footnotes,
  page_header,
  page_footer,
  column_labels
)
}
}
