% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcon02.R
\name{create_table_02}
\alias{create_table_02}
\title{Create Demographic Table}
\usage{
create_table_02(
  df,
  vars = c("AGE", "SEX", "ETHNICC", "HEIGHTBL", "WEIGHTBL", "BMIBL"),
  lbl_vars,
  .labels
)
}
\arguments{
\item{df}{The input dataframe containing the raw data.}

\item{vars}{A character vector of variable names. Default is c("AGE", "SEX", "ETHNICC", "HEIGHTBL", "WEIGHTBL","BMIBL").}

\item{lbl_vars}{A named character vector of variable labels. Default is c(AGE = "年龄（岁）", SEX = "性别, n", ETHNICC = "民族, n", HEIGHTBL = "身高（cm）", WEIGHTBL = "体重（kg）", BMIBL = "基线体重指数（kg/m2)").}

\item{.labels}{A named character vector of labels for the statistics. Default is c(n = "例数 (缺失)", mean_sd = "均值(标准差)", median = "中位数", quantiles = "Q1. Q3", range = "最小值. 最大值", count_fraction = "n").}
}
\value{
A dataframe that represents the demographic table.
}
\description{
This function creates a demographic table for clinical trial reports, often referred to as Table 02.
It uses the `make_table_02_huarun` function, which generates a table with specified variables and labels. 
The function also modifies the 'V1' column in the resulting table based on the labels passed.
}
\examples{
\dontrun{
vars <- c("AGE", "SEX", "ETHNICC")
demographic_table <- create_table_02(df, vars = vars)
}
}
