% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcon03.R
\name{create_table_03}
\alias{create_table_03}
\title{Create Disposition Table}
\usage{
create_table_03(adsl)
}
\arguments{
\item{adsl}{The ADSL dataframe containing the subject-level analysis data.}
}
\value{
A dataframe that represents the disposition table.
}
\description{
This function creates a disposition table for clinical trial reports, often referred to as Table 03.
It uses the `make_table_03` function from the `falcon` package, which is not exported and therefore accessed directly.
}
\examples{
\dontrun{
disposition_table <- create_table_03(adsl)
}
}
