% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcon22.R
\name{create_table_22}
\alias{create_table_22}
\title{Create Table 22}
\usage{
create_table_22(df, adsl)
}
\arguments{
\item{df}{The input dataframe.}

\item{adsl}{The ADSL dataframe.}
}
\value{
A dataframe that represents Table 22.
}
\description{
This function creates Table 22, which is a demographic table in clinical trial reports.
}
\examples{
\dontrun{
table_22 <- create_table_22(df, adsl)
}
}
