% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcon22.R
\name{load_libraries_and_data}
\alias{load_libraries_and_data}
\title{Load libraries and data}
\usage{
load_libraries_and_data()
}
\value{
A list containing the dataframe for analysis and the original adsl dataframe.
}
\description{
This function loads the necessary libraries and data for the analysis.
}
\examples{
\dontrun{
data <- load_libraries_and_data()
}
}
