% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcon03.R
\name{load_libraries_and_data_03}
\alias{load_libraries_and_data_03}
\title{Load libraries and preprocess data for Falcon 03}
\usage{
load_libraries_and_data_03()
}
\value{
A dataframe named adsl
}
\description{
This function loads the necessary libraries and preprocesses the ADSL data for the Falcon 03 analysis.
The preprocessing includes setting a seed for reproducibility, adding missing values to the RANDDT column,
and adding new variables (ENRLDT, SCRNFL, SCRNFRS, SCRNFAILFL) to the dataframe.
}
