% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcon02.R
\name{make_table_02_huarun}
\alias{make_table_02_huarun}
\title{Make Table 02}
\usage{
make_table_02_huarun(
  df,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  arm_var = "ARM",
  saffl_var = "SAFFL",
  vars = c("AGE", "HEIGHTBL", "WEIGHTBL", "BMIBL"),
  vars_no_n = c("SEX", "ETHNICC"),
  lbl_vars = formatters::var_labels(df, fill = TRUE)[c(vars, vars_no_n)],
  lbl_overall = "Total Population",
  na_rm = FALSE,
  prune_0 = TRUE,
  annotations = NULL,
  .labels = c(n = "Number (Missing)", mean_sd = "Mean (SD)", median = "Median", quantiles
    = c("Q1. Q3"), range = c("Min. Max"), count_fraction = "n (\%)")
)
}
\arguments{
\item{df}{A data frame.}

\item{alt_counts_df}{An alternative counts data frame. Default is NULL.}

\item{show_colcounts}{Logical. If TRUE, column counts are shown. Default is TRUE.}

\item{arm_var}{The arm variable. Default is "ARM".}

\item{saffl_var}{The saffl variable. Default is "SAFFL".}

\item{vars}{A character vector of variable names. Default is c("AGE", "HEIGHTBL", "WEIGHTBL","BMIBL").}

\item{vars_no_n}{A character vector of variable names without 'n'. Default is c("SEX", "ETHNICC").}

\item{lbl_vars}{A named character vector of variable labels. Default is formatters::var_labels(df, fill = TRUE)[c(vars, vars_no_n)].}

\item{lbl_overall}{The label for overall. Default is "Total Population".}

\item{na_rm}{Logical. If TRUE, NA values are removed. Default is FALSE.}

\item{prune_0}{Logical. If TRUE, 0 values are pruned. Default is TRUE.}

\item{annotations}{Annotations for the table. Default is NULL.}

\item{.labels}{A named character vector of labels for the statistics. Default is c(n = "Number (Missing)", mean_sd = "Mean (SD)", median = "Median", quantiles = c("Q1. Q3"), range =c("Min. Max"), count_fraction = "n").}
}
\value{
A table.
}
\description{
This function creates a table with specified variables and labels. The labels are passed as an argument to the function. The function also modifies the 'V1' column in the resulting table based on the labels passed.
}
\details{
* `df` must contain the variables specified by `vars`, `arm_var`, and `saffl_var`.
* If specified, `alt_counts_df` must contain the variables specified by `arm_var`, `id_var`, and `saffl_var`.
* Flag variables (i.e. `XXXFL`) are expected to have two levels: `"Y"` (true) and `"N"` (false). Missing values in
  flag variables are treated as `"N"`.
* Columns are split by arm. Overall population column is included by default (see `lbl_overall` argument).
* Information from either ADSUB or ADVS is generally included into `df` prior to analysis.
* Numbers in table for non-numeric variables represent the absolute numbers of patients and fraction of `N`.
* All-zero rows are removed by default (see `prune_0` argument).
}
\section{Functions}{
\itemize{
\item \code{make_table_02_huarun()}: Create FDA table 2 using functions from `rtables` and `tern`.

}}
\examples{
\dontrun{
vars <- c("AGE", "SEX", "ETHNICC")
.labels <- c(n = "例数 (缺失)", mean_sd = "均值(标准差)")
make_table_02_huarun(df = data.frame(), vars = vars, .labels = .labels)
}
}
