% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falcon22.R
\name{make_table_22}
\alias{make_table_22}
\title{Create a table for a given dataframe}
\usage{
make_table_22(
  df,
  alt_counts_df = NULL,
  show_colcounts = TRUE,
  arm_var = "ARM",
  saffl_var = "SAFFL",
  vars = c("SEX", "AGEGR1", "RACE", "ETHNIC"),
  denom = c("N_s", "N_col", "n"),
  lbl_overall = NULL,
  lbl_vars = formatters::var_labels(df, fill = TRUE)[vars],
  prune_0 = FALSE,
  annotations = NULL
)
}
\arguments{
\item{df}{The input dataframe.}

\item{alt_counts_df}{An optional dataframe for alternative counts. Default is NULL.}

\item{show_colcounts}{A logical indicating whether to show column counts. Default is TRUE.}

\item{arm_var}{The variable for arm. Default is "ARM".}

\item{saffl_var}{The variable for saffl. Default is "SAFFL".}

\item{vars}{A character vector of variables. Default is c("SEX", "AGEGR1", "RACE", "ETHNIC").}

\item{denom}{A character vector for denominator. Default is c("N_s", "N_col", "n").}

\item{lbl_overall}{An optional label for overall. Default is NULL.}

\item{lbl_vars}{A character vector of labels for variables. Default is the labels of vars in df.}

\item{prune_0}{A logical indicating whether to prune 0. Default is FALSE.}

\item{annotations}{An optional annotations. Default is NULL.}
}
\value{
A table based on the input dataframe and other parameters.
}
\description{
This function creates a table based on the input dataframe and other parameters.
}
\examples{
\dontrun{
make_table_22(df)
}
}
